stabilisation en attente de validation v1.

+++ app/Http/Controllers/EspeceController.php
@@ -2,61 +2,47 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Espece;
+use App\Models\Crime;
 use App\Models\Ordre;
+use App\Models\Espece;
+use App\Models\Restriction;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
-use Illuminate\Support\Str;
 
 class EspeceController extends Controller
 {
+
+
+
     public function __construct()
     {
         $this->middleware('auth');
     }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         $especes=Espece::orderBy('nom','asc')->get();
 
         return view('pages.backoffice.especes.index',compact('especes'));
     }
-
-
     public function create()
     {
-        
-        return view('pages.backoffice.especes.form',[
-            'ordres' =>Ordre::all()
+        return view('pages.backoffice.especes.createdit', [
+            'espece'                      =>  new Espece(),            
+            'ordres'                      =>  Ordre::all(),
+            'titrePage'                   => "Ajouter une nouvelle espèce",
+            'btnAction'                   => "Ajouter"
         ]);
     }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {   
-
-
-      
         $data=request()->validate([
-            'nom'=> ['required','string','max:255','min:3'],
+            'nom'=> ['required','string','max:255','min:3','unique:especes'],
             'famille'=> ['required','string','max:255','min:3'],
             'type'=> ['required','string','max:255'],
             'ordre_id'=> ['required','integer'],
             'nom_scientifique'=> ['required','string','max:255','min:3'],
             'photo'=> ['image'],
-
-
-
-
           ]);
             $espece= new Espece;
           if($request->hasFile('photo')){
@@ -76,40 +62,22 @@ class EspeceController extends Controller
           $request->session()->flash('status', 'Espèce ajoutée avec succès');
           return redirect()->route('especes.index');
     }
-
-
     public function show($uuid)
     {
-        $espece=Espece::where('uuid',$uuid)->first();
-
-        return view('pages.backoffice.especes.show', compact('espece'));
+        return view('pages.backoffice.especes.show',[
+            'espece'   => Espece::where('uuid',$uuid)->first(),
+            'crimes' => Crime::where('espece_id',Espece::where('uuid',$uuid)->first()->id )->get()
+        ]);
     }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Espece  $unite
-     * @return \Illuminate\Http\Response
-     */
     public function edit($uuid)
     {
-
-        $espece=Espece::where('uuid',$uuid)->first();
-
-
-        return view('pages.backoffice.especes.edit', [
-            'ordres' =>Ordre::all(),
-            'espece'=>Espece::where('uuid',$uuid)->first()
+            return view('pages.backoffice.especes.createdit', [
+                'espece'            =>Espece::where('uuid',$uuid)->first(),            
+                'ordres'            =>  Ordre::all(),
+                'titrePage'         => "Mise a jours de ".Espece::where('uuid',$uuid)->first()->nom,
+                'btnAction'         => "Mettre a jours"
             ]);
     }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Espece  $unite
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $uuid)
     {
         $data=request()->validate([
@@ -138,20 +106,18 @@ class EspeceController extends Controller
           $request->session()->flash('status', 'Espèce mise a jours avec succès');
           return redirect()->route('especes.index');
     }
-
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Espece  $unite
-     * @return \Illuminate\Http\Response
-     */
-
     public function destroy(Request $request,$uuid)
     {
         $espece=Espece::where('uuid',$uuid)->first();
-        $espece->delete();
-
-        return redirect()->route('especes.index')->with('status','Espece supprimée avec succès');
+        $restriction = new Restriction;
+        $restrictions = $restriction->check($espece->id,[
+            ['foreignkey'=>'espece_id','modelname'=>'crime'],
+        ]);
+        if ($restrictions){
+        return redirect()->back()->with('danger',$restrictions['message']);
+        } else {
+            $espece->delete();
+            return redirect()->route('especes.index') ->with('status','Espece supprimée avec succès');
+        }
     }
 }
